Updating deployed pods according to CR spec
Now changed only container resources and number of containers in pod

Also made some refactoring
