+++ cmd/percona-xtradb-cluster-operator/main.go
@@ -5,14 +5,14 @@ import (
 	"runtime"
 	"time"
 
-	stub "github.com/Percona-Lab/percona-xtradb-cluster-operator/pkg/stub"
-	"github.com/Percona-Lab/percona-xtradb-cluster-operator/version"
 	sdk "github.com/operator-framework/operator-sdk/pkg/sdk"
 	k8sutil "github.com/operator-framework/operator-sdk/pkg/util/k8sutil"
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
-
 	"github.com/sirupsen/logrus"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
+
+	"github.com/Percona-Lab/percona-xtradb-cluster-operator/pkg/pxc"
+	"github.com/Percona-Lab/percona-xtradb-cluster-operator/version"
 )
 
 func printVersion() {
@@ -42,6 +42,6 @@ func main() {
 	resyncPeriod := 5 * time.Second
 	logrus.Infof("Watching %s, %s, %s, %v", resource, kind, namespace, resyncPeriod)
 	sdk.Watch(resource, kind, namespace, resyncPeriod)
-	sdk.Handle(stub.NewHandler(*sv))
+	sdk.Handle(pxc.New(*sv))
 	sdk.Run(context.TODO())
 }
